testing: merge code/test_terminal_writer.py into io/test_terminalwriter.py
